LPS-111200 Merge provider and provider factory
since provider can only be created by its factory
