+++ modules/apps/portal-crypto-hash/portal-crypto-hash-provider/portal-crypto-hash-provider-bcrypt/src/main/java/com/liferay/portal/crypto/hash/provider/bcrypt/internal/BCryptCryptoHashProviderFactory.java
@@ -16,8 +16,14 @@ package com.liferay.portal.crypto.hash.provider.bcrypt.internal;
 
 import com.liferay.portal.crypto.hash.spi.CryptoHashProvider;
 import com.liferay.portal.crypto.hash.spi.CryptoHashProviderFactory;
+import com.liferay.portal.crypto.hash.spi.CryptoHashProviderResponse;
+import com.liferay.portal.kernel.util.MapUtil;
+
+import java.nio.charset.StandardCharsets;
 
 import java.util.Map;
+
+import jodd.crypt.BCrypt;
 
 /**
  * @author Arthur Chan
@@ -47,5 +53,50 @@ public class BCryptCryptoHashProviderFactory
 	}
 
 	private static final String _CRYPTO_HASH_PROVIDER_FACTORY_NAME = "BCrypt";
+
+	private static class BCryptCryptoHashProvider
+		implements CryptoHashProvider {
+
+		public static final String ROUNDS = "rounds";
+
+		public BCryptCryptoHashProvider(String cryptoHashProviderFactoryName) {
+			_cryptoHashProviderFactoryName = cryptoHashProviderFactoryName;
+
+			_rounds = _DEFAULT_ROUNDS;
+		}
+
+		public BCryptCryptoHashProvider(
+			String cryptoHashProviderFactoryName,
+			Map<String, ?> cryptoHashProviderProperties) {
+
+			_cryptoHashProviderFactoryName = cryptoHashProviderFactoryName;
+
+			_rounds = MapUtil.getInteger(cryptoHashProviderProperties, ROUNDS);
+		}
+
+		@Override
+		public CryptoHashProviderResponse generate(byte[] salt, byte[] input) {
+			String hashedPassword = BCrypt.hashpw(
+				new String(input, StandardCharsets.US_ASCII),
+				new String(salt, StandardCharsets.US_ASCII));
+
+			return new CryptoHashProviderResponse(
+				_cryptoHashProviderFactoryName, null,
+				hashedPassword.getBytes(StandardCharsets.US_ASCII));
+		}
+
+		@Override
+		public byte[] generateSalt() {
+			String salt = BCrypt.gensalt(_rounds);
+
+			return salt.getBytes(StandardCharsets.US_ASCII);
+		}
+
+		private static final int _DEFAULT_ROUNDS = 10;
+
+		private final String _cryptoHashProviderFactoryName;
+		private final int _rounds;
+
+	}
 
 }
